Add files via upload
✅ DAY 7Project 1: Even or Odd Checker
📌 Extended Description:
On Day 7 of my #100DaysOfPython journey, I worked on creating an Even or Odd Checker. This small but essential program accepts an integer input from the user and determines whether it is even or odd. Behind the scenes, it uses the modulus operator %, which returns the remainder of division. If a number is divisible by 2 (i.e., the remainder is zero), it is even; otherwise, it is odd.

This project helped me understand how conditional statements (if-else) work in Python. It’s also a great example of how we can use functions to make code reusable and organized.

Through this, I reinforced:

Reading input from the user using input()

Type conversion from string to integer using int()

Using conditional logic

Returning values from functions

This was a simple yet powerful exercise that forms the building block of logic-building in programming.