Add files via upload
🚀 Day 94 – Integer to Roman Conversion
🔹 Problem Statement

We are given an integer number num. The task is to convert it into a Roman numeral string based on the following rules:

Roman numerals are made by combining symbols:

I = 1, V = 5, X = 10, L = 50, C = 100, D = 500, M = 1000


Conversion rules:

Symbols are appended from highest to lowest.

If the number starts with 4 or 9, use the subtractive form (like IV for 4, IX for 9, XL for 40, etc.).

Only powers of 10 symbols (I, X, C, M) can be repeated at most 3 times in a row.

For 4 consecutive symbols, we switch to subtractive form.